@@ -12,5 +12,5 @@
 // Generated on: 2016.09.21 at 01:37:30 PM CEST 
 //
 
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://www.sa.dk/xmlns/diark/1.0", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+@jakarta.xml.bind.annotation.XmlSchema(namespace = "http://www.sa.dk/xmlns/diark/1.0", elementFormDefault = jakarta.xml.bind.annotation.XmlNsForm.QUALIFIED)
 package dk.sa.xmlns.diark._1_0.fileindex;
