@@ -159,7 +159,16 @@ public class PostgreSQLJDBCDatatypeImporter extends JDBCDatatypeImporter {
     // The maximum allowed precision when explicitly specified in the type
     // declaration is 1000, so if we find more than that it means that this type was
     // declared without precision nor scale
-    if (columnSize > 1000) {
+
+    //20240626 alindo: when upgrading the conector to version 42.7.3 it sets the data type
+    // scale and precision to 0 when they are not specified which is against SIARD specification
+    // https://github.com/pgjdbc/pgjdbc/issues/2188
+    if (columnSize == 0) {
+      type.setSql99TypeName("NUMERIC", NUMERIC_MAX_PRECISION_NUMBER, NUMERIC_MAX_SCALE_NUMBER);
+      type.setSql2008TypeName("NUMERIC", NUMERIC_MAX_PRECISION_NUMBER, NUMERIC_MAX_SCALE_NUMBER);
+      reporter.customMessage(this.getClass().getName(),
+        "Column data length is 0. Replacing the length to the max data length " + NUMERIC_MAX_PRECISION_NUMBER);
+    } else if (columnSize > 1000) {
       type.setSql99TypeName("NUMERIC", NUMERIC_MAX_PRECISION_NUMBER, NUMERIC_MAX_SCALE_NUMBER);
       type.setSql2008TypeName("NUMERIC", NUMERIC_MAX_PRECISION_NUMBER, NUMERIC_MAX_SCALE_NUMBER);
       reporter.customMessage(this.getClass().getName(),
